Wheels, tracks etc updated with group ID's
Group 0 is special, it ONLY modifies the currently selected part and
other parts will not affect its settings. Using the Apply Settings
button will apply settings to other parts in the craft with the same
group number. Steering ratio is calculated only for parts with the same
ID, so if you have a rover with a sib rover you don't want the ratio
calculation to apply to (it's docked but not on the floor or in a
different orientation), set it to anther group to the main one. Love it,
so elegantly simple.