Major update
Added loads of wheel features, some of which I've not even tested yet.
Try out the retract feature on the large and medium wheels, as well as
the ability to change the ride height on all wheels. Repulsor water
hovering will be broken until further notice, and repulsor wheels are
still not in the middle of a re-write.