A large integer value of changes.
So, I've done a little reworking of the scaletype definitions and
implemented a more flexible check within the part configs for those
using MM to remove the module if the TweakScale plugin is not found,
reducing errors in the log when the parts are compiled.

A few unnecessary spaces and blank lines have been removed and some
redundant module definitions have been commented out (just in case they
still have a purpose that I am unaware of) such as for the part icon
fixer, which went config-less quite some time ago.

Finally, the scaletypes got their first addition which will eventually
interact with the new KFDustFX, which I am integrating from my DustFX
project (based on, but completely rewritten from, CollisionFX) and have
commented out a few references to CollisionFX in the part configs in
preparation for the new module which will handle all visual elements of
the interaction between rolling support systems (wheels) or similar
parts.  CollisionFX can be used later for audio collisions unless we get
a chance to implement that into KF too.  That's all for now.