Lots of changes... see description.
Okay, so I've cleaned up the moved source files and put them back where
they belong.  I've done a lot of reworking on the DustFX and I've come
to the conclusion that, since performance could possibly be reduced by
large amounts of dust, that we need to make it at optional as possible.
Until I figure out how to add a config option to enable/disable the
effects on a global level, or add a GUI option for global/per-part
settings which can be persisted between game loads, for the time being
the effects will be dependant on the user having MM installed.  This
also allows me to make sure that CollisionFX (not my mod, so no control
over it) won't have its module on the part messing everything up.  The
patch I made removes any instances of that module from the parts before
it adds the DustFX portion.  Work will continue towards a better option
in the mean time.

I've cleaned up some of the part files of commented-out redundant
modules, and I located a single mistake where the old module name
"Repulsor" was used when it is now, in fact, "KFRepulsor" and would thus
no longer work properly.

Scaletypes have had a redo in several places as well, but nothing needs
to be done in the parts to make it work.  I'm still not up to date on
the new "free scale for all" methods, but that will have to wait for
another time.  I did have one issue recently where TS stopped showing up
for specific parts in the KF stock wheel set, but I think I have that
squashed.

Re-added the part info entries into the other modules, all of which (I
think?) are KSPFields which means you could, potentially, override the
string from the part configs.  I have yet to try this in practice, but I
see no problem with that functionality as it is now.  We may never need
it, but it might be handy.