Just some tweaks here and there.
A few changes to  a few source files that were annoying me (even if you
consider I don't edit them... why must these things haunt me so?)

Re-enabled the OnCollisionEnter method in DustFX because it is used for
an untested wheel impact sound.  Added a field definition to another of
the part info strings, so it can be overridden in the part config if
need arises.  Updated the DustFX patch to account for the screwdrive
needing a lower velocity magnitude due to the extra time it spends in
contact with the ground.  I think that sums it up.

Oh, and the log utils got more updates, but the utility class is still
not ready to be used in any form or another at this time.  Work is slow,
but I'll finish eventually and it will be fabulous!  Or it'll kill me,
but we're gonna aim for "Fabulous!"