basic cleanup and a scaletype fixer
I noticed that we switched to 2-value floatcurves instead of those
4-value ones for most of the zero-sporting last two values, so I culled
them up a bit where appropriate.  I also added a MM patch to the end of
the scaletype definition file which will fix any parts that remain which
use the old generic KFTrack scaletype which is now called KFTrackMedium.