DustColorDefinitions stuff.
I cleaned up and remerged the data from the active config.  I would not
suggest running with both configs active since we do not load the
definitions from a specific file, but rather from the game database.  In
theory, the one loaded last will override the other.

I cleaned up the file too.  "DefaultColor" only needs to exist on the
global level, and the planetary level.  "Color" only needs to exist on
the biome-specific level.  I also removed my attempt to add a "water"
biome specification from the files which may have messed up some things
somewhere, except of Kerbin where it is an official biome specification.
For now, if the water spray works at all, it'll have to be a hard coded
color value.  I'm working on reworking the config system.