DDSified the textures...
at least, most of them.  It turns out the DXT5nm format, which is what
also turns the colors into a pinkish hue, also removes some details in
certain normal maps.  In the ones that lost too many details, I kept
them in PNG format.  Only the rover body texture has mip maps currently
due to its size, but if it looks too ugly I can always revert to a
version that is resized to be 1024x1024, but without mipmaps, which
looks good on my end.