Made some tweaks, then got hit by a github merge error.  Fixed now.
Repaired the DustFX entries in the configs, removed a space from some of
the part titles, gave KFGlobals a new entry that will be in the code in
a few minutes, and added a new sound for the repulsor which is basically
just an edit of the original that I found to be more interesting.  I
have not configured any parts to use it however.  Waiting for lo-fi to
approve it.